nl_l3: use the new std::set<nh_stub> -> ecmp group id
Now that we have a way of mapping nexthops sets to a ecmp group, use
this instead of mapping a set of l3 interface ids, so we have stable
ecmp groups regardless of availability of nexthops.

This greatly simplifies nl_l3::del_l3_unicast_route() since we now can
remove the ecmp group before attempting to resolve the nh_stubs to
neighs, allowing us to directly drop the l3 interface references in one
go.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>