allow configuring the default untagged VLAN for ports
Using unbridged, untagged ports at the same time as VID 1 as default
PVID on a bridge may cause unexpected leakage and forwarding of traffic.

Moving to the reserved VID 4095 would avoid this, but we cannot
unconditionally use it until we solved #414, as otherwise we would leak
packets with VID 4095.

So for now allow changing the VID used internally for ports, but keep
the default at 1.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>