adapt mdb handling to fixed upstream working
With a recent fix, libnl will now send the updated object instead of the
update as new for callback_v2(). Unfortuntely the mdb code relied on
getting incremental updates as new, so this change broke mdb handling.

To fix this, rework the mdb handling the following:

1. convert all entries of old and new into seperate, ordered sets
2. for any entry in the old, but not new leave the group
3. for any entry in the new, but not old join the group

This with empty sets for old respective new this can also be used new
and delete of the object.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>