nl_l3::is_l3_neigh_routable(): fix routability check
L3 neighbours are always directly connected, so they need to be on link
scoped routes. Having a nexthop with an ip is actually the opposite of
that.

So change the check to check the route scope, and only mark neighbours
as routable if the route is link scoped (i.e. no gw/via nexthops).

Fixes neighbours being mistakenly assumed as routable if there is a
gateway route present anywhere that routes that ip address (e.g. a
default route in the management interface).

Also remove a stray semicolon at the end of the function.

Fixes: 0fdba0d69587 ("nl_l3: only route l3 neighs if we have a route for them")
Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>