nl_bridge: add support for updating ageing time
Add support for updating the ageing time by updating all table 50 flows
with a new soft idle timeout.

This works by setting a new default idle timeout value, then triggering
an update for all tracked fdb entries.

This will cause a reset of the entry's age to 0, but it's the closest we
can do within the OpenFlow framework.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>