of-dpa: remove micro-optimization for l3 interface ids
Remove the micro-optimization for controller::l3_egress_remove((), as it
would have only done anything if the removed id was the last one added.

On top of that the condition was broken, so it never even triggered, as
it checks for the freed id to be one more than the next id to use, which
will never be true.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>