set rx rate to unlimited for knet
Since KNET uses a different channel for packets than OpenFlow and is
much more efficient, higher packet rates than 1024 (the default) is
easily doable.

So set the rx rate to unlimited by default when using KNET, but leave it
at 1024 pps for TAP.

In case this causes issue, add a flag to set a custom limit.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>