cnetlink: allow looking up routes by dst and nhid
Allow looking up routes by their nhid instead of a nexthop directly.
So allow this to work, add a nhid field to nh_stub so we can store the
nhid of route nexthops.

Note that a route will always point to one nhid, which may either be a
single nexthop, or a nexthop group. Therefore we need to store the nhid
of the route, not any potential nhids or nexthops in the nh_stub.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>