cnetlink: properly handle ll neigh handable changes
With ae62117d337c ("cnetlink: pass on layer 2 neigh updates"), we make
sure to handle updates only on neighs we support, so we check both old
and new to be usable.

But this misses the transitions where a ll neigh becomes usable due to
updates or vice versa.

Fix this by treating those changes as a new or deleted ll neigh insted
of ignoring it alltogether.

Fixes: ae62117d337c ("cnetlink: pass on layer 2 neigh updates")
Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>