controller: fix setting station move configuration
In controller::port_set_move_learn(), we convert the passed enum to
flags, but then pass on the original enum instead to OF-DPA. This sets a
random configuration, which may or may not be what we actually
intentend.

Fix this by passing the correct flags variable.

Fixes: 87b00ee6b64e ("ofdpa_client: expose per port learning configuration")
Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>
(cherry picked from commit 00ed2ca2f4a0960031dec4eaaf0f1f2682dca383)
Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>