knet_manager: do not try to set state for non-existing links
Since OF-DPA brings all links up at boot, there is a chance that the is
a link state change before we received the port table and created the
ports.

If this happens, then we get a change modification for speed/link, but
there is no link to modify. This is harmless since we will set the link
to down once we do create the port interface, but triggers a warning
from knet on startup:

kernel: linux-bcm-knet (1197): Warning: unknown network interface: 'port1'
kernel: linux-bcm-knet (1197): Warning: unknown network interface: 'port2'
...
kernel: linux-bcm-knet (1197): Warning: unknown network interface: 'port54'

So check if the link is created before trying to update its state.

Fixes: 40f68c81c10a ("add knet manager for using knet based interfaces")
Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>