nl_vlan: handle copying bridge vlan config
Instead of requiring the bond module to know how we manage vlans for
bridges, let nl_vlan do it now that we track the bridge vlan config as
well.

We still need to notify the bridge module though for pre-existing bonds
that are attached to the bridge, so do that when creating the bond
mapping.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>