linter: invert the push branch restrictions
Instead of ignoring master/main and linting everything eslse, lint on
main and any presumed stable branch (that starts with [1-9].) only.

If this is a PR branch, the pull_request action should already trigger
linting for it, and if there is no PR for it, then code in the branch is
likely not ready yet for prime time, so it does not need to conform to
the code standard yet.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>