nl_l3: ignore updates on link scoped routes
If we get an update on a route, we first "add" the route again, then
"delete" the old version in update route mode which leaves the flow, but
updates everything else.

This works fine for routes with ip nexthops, but for link local rotes,
we would first mark all l3 neighbors as routable, then on handling the
delete, mark them all unroutable, breaking any routes via nexthops on
that interface.

So ignore updates for link local routes, since we cannot do
anything meaningful with the update anyway.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>