Revert "cnetlink: sync route cache on linkdown events"
Syncing routes on link down may trigger updates on link local routes due
to route flag changes, but baseboxd is not prepared to handle this, and
causes all l3 neighbors on that link to be treated as unroutable.

IPv6 routes with nexthops seem to behave differently than IPv4 routes,
and seem to reappear silently on link up again. Since there is no
notification (again) from the kernel, we miss the reappearance of the
route.

Due to the amount of breakage, lets's revert this for now until we have
addresses these two issues.

This reverts commit 9a58b5ddbe9e7ce65228f4bc5bcf325e6905e14c.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>