nl_vlan: simplify ::get_vid()
We really only care about two cases, is this a VLAN interface or not, so
reduce the check to just that.

In theory we could also try to find out the bridge PVID for the raw
bridge link, but that would require updating flows and groups on PVID
changes, which we currently don't support.

This also avoids a crash with VRFs attached to raw interfaces, which is
a valid thing to do, but our code currently assumes that this is only
done for VLAN interfaces, causing an assertion failure when trying to
get the VLAN id of the assumed base link.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>