of-dpa: set idle timeout on the flowmod directly
Instead of setting, then unsetting the global idle timeout setting in
rofl-ofdpa, set the idle timeout directly in the generated flowmod.

This also means we need to enable flow removal notifications ourselves.

This avoids that the idle timeout setting remains set at the non-default
value if for any reason sending the flow mod creates an exception.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>