cnetlink: let ::get_vlan_links() take references
We need to make sure that any objects we retrieve from the cache via
nl_cache_foreach(_filter) have an explicit reference, else they can get
released and free'd while we use them, leading to use-after-free and
accessing reused memory.

So change cnetlink::get_vlan_links() to take a reference on all returned
objects.

This also means that we need to explicitly drop the references again
when we are finished, so do that as well for all callers.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>