nl_l3: let ::search_neigh_cache() just return a bool
We never do anything with the found neigh list by
nl_l3::search_neigh_cache() and just check if there is one, so we might
as well directly return the result from nl_l3::search_neigh_cache().

By directly returning a bool, we avoid the issue of (not) having a
reference of the neighs found.

While at it, rename it to nl_l3::have_neigh() to more closely match what it
now does.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>