nl_vxlan: let ::create_endpoint() take references during use
We need to make sure that any objects we retrieve from the cache via
nl_cache_foreach(_filter) have an explicit reference, else they can get
released and free'd while we use them, leading to use-after-free and
accessing reused memory.

So change nl_vxlan::create_endpoint() to take a reference on all objects
retrieved from the cache.

This also means that we need to explicitly drop the references again
when we are finished.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>