cnetlink: process netlink events synchronously
Currently we switch between collecting netlink events and processing them.
This makes the processing bursty, and since libnl will have finished
applying them to the cache it means the cache will be at a state after
applying the last event of the current batch.

Change this to process each netlink event as it comes in instead. This
ensures that the current libnl caches are in the state of when that
event happened. This means that we do not need to look into the future
for anymore for objects that get deleted in the same batch.

Since we do not queue messages anymore, we need to make sure to be able
to handle netlink notifications before we start creating port
interfaces, so notify about switch being up earlier and change the state
to running.

Signed-off-by: Jonas Gorski <jonas.gorski@bisdn.de>